change namespace and package names
This changes the build/auto versioning process to using gitversion and having SixLabors.ImageSharp & SixLabors.ImageSharp.Drawing both have the same version number.
When development on one half or the other starts to change cadence and need to drift out of sync then we can just move Drawing to its own repo and manage it as a
separate component like the web/middlewear pieces will be like.